13th Jan 2013. Implements GA (with inspyred) on curiosity loops
A dummy fitness function appears in motorBabbling9.py. However, the sm
matrix specified by each loop is saved at the end of each short fitness
assessment, and this time-series can be used as the basis for
calculating the fitness of that loop.

Only one motor is controlled at any time, and only 2 sensors influence
that motor via the FFNN.

Currently the variation operator is null, so there is no mutation or
crossover.