A 2 atom action molecule is hand-coded.
Atom 1. Adds two sensor inputs and writes to ALMemory
Atom 2. The Hiil Climber Atom takes this ALMemory value as fitness, and
executes a changeAngle to a set of angles that are the parameters.
Parameter assessment takes place EACH timestep, so the system tracks
instentaneous improvement in fitness.

Currently the sensory inputs to the function atom are two accelerometer
values and the motor output of the Hill climber atom is a randomly
chosen motor effector list.